Encapsulate select() calls, part 1
The problem: select() can only handle a fixed number of file descriptors,
1024 on most systems, and it has a very awkward low-level interface.  We
will remove this restriction and clean up the related code in a series of
commits.  Each commit will intentionally be small, but self-contained, easy
to understand, and most of the time mechanical.  The whole series is
intended to be an example for improving an aspect of out SW in a structured
simple way, without the need for huge commits or great display of ingenuity.

This commit at hand only does something extremely simple and is a totally
mechanical refactoring: Put the select() call into a class, providing a 1:1
interface for now.  Doesn't do very much in itself, but it is a starting
point for things to come...

Change-Id: I048ca6e7e2bbde7c3c4e2951baaef29874ed72c6