Encapsulate select() calls, part 3
Use C++'s standard std::chrono facilities instead of fiddling around with
the low-level timeval struct.  This is a mostly mechanic change, with the
exception of icmpsender's main_loop: This relied on select() modifying the
timeout, so we had to measure the duration for ourselves.  This way we can
keep Poller::poll()'s interface simple, and main_loop() needs some massaging
in the future, anyway.

Note that we can reduce our dependency on ChronoUtils.h quite a bit now,
which is a nice side-effect of this change.

Change-Id: I7ca388dd8cef4507d1f18df6a858bd84cbaaa790