Encapsulate select() calls, part 11
Generalize the interface a bit: Instead of setting/testing a single kind of
event at a time, use a set of polling events.  The former interface was OK
for the underlying select()-based implementation, but it was inadequate for
a future poll()-based implementation.

To use the involved bit masks in a type-safe way, a separate header
containing templated operators is introduced.  This seems to be overkill for
the task at hand, but we can nicely use this at other places later.

Alas, this commit contains some tiny fixes to make cppcheck and clang-tidy
happy again.  There seem to be some heuristics which have been triggered
accidentally, so let's just fix them now.

Change-Id: I29e788d8fb763778d1125f1a1a5543a4e97090aa