4989 Added handling of host downtimes during event creation
The Event Console is now saving whether or not a host that can be found in the core
was in downtime or not. Based on this information the event handling can be influenced
in different ways.

The idea is: In the moment a host is in downtime all events generated by the host are
"handled problems". The events should be created and be visible during the downtime to
assist during maintenance. After maintenance the events that have been created in the
downtime are not interesting anymore.

What has changed:

<ul>
<li>Each event has a new attribute <tt>event_host_in_downtime</tt>. Assuming <i>0</i>
(no) for existing events.</li>
<li>It is possible to filter event views by this information</li>
<li>Separate counting events are created for in downtime / not in downtime</li>
<li>Events are added during downtime and visible in the default events views<li>
<li>The tactical overview does not count events created in host downtime in the
"handled problems" column</li>
<li>Events created during downtime will be archived automatically after the host
left the downtime</li>
<li>Actions are handled independent of the downtime by default. But you have the
option to only execute actions when the host was not in downtime</li>
</ul>

Events that are not associated with a monitoring host in the local site are
not affected by the changes mentioned above.

Change-Id: Ic571b8488134095c019965cc76ea36efcbfdec67