4563 FIX Fix issue with @ in metric title
The @ sign is used to append a custom unit when used within metric titles.
This prevents graphs from being displayed when an @ without a special meaning
appears within a metric title. The graphing system now checks for an
existing unit within the metric title amd leaves the title untouched otherwise.

Change-Id: Ic7657bd01b779385851a7a2bfac4e9c27285c685