4993 FIX Fixed visibility of events for users with limited access to events
When a user is not permitted to see all events and can only see the events
of the host he/she is a contact for and maybe also events not related to
a host existing in the monitoring core the tactical overview snapin counted
the numbers of open events not correctly.

The event list view was filtering wrong in case a user has the "see all"
permission and not the "see unrelated" permission.

Change-Id: I475a6adfcf8501feafc9f0c2064de477cefe9203