4811 mssql_backup: Set limits for different type of backups
For the check mssql_backup it is now possible to set limits for
different types of backups (e.g. database backup, log backup etc.)
instead of just one limit that is valid for all backups. The
limits can be set in the rule "MSSQL Time since last Backup".
If no backup types are supported by the agent the parameter
"Database Backup" can be used to set a general limit.

Change-Id: Iea7ff7a86d03e6fd5a043c9bdb6f120053febed8