5003 FIX Fixed missing filtering by effective contact groups of events
Since 1.4 the Event console has the concept of "effective contact groups" for
an event. The contact groups of an event can either be set by a matched host
or by the matching rule. Or by both. For the later case there is an option
to configure the precedence in the rule.

The GUI was not correctly reflecting this feature because there was no painter
and no filter for the effective contact groups.

Change-Id: I81b0285f3f9b0dff8326ad58eabaf4c570e97a50