5039 Livestatus proxy: Connection parameters can now be configured globally
The connection parameters of the Livestatus Proxy can now be configured
globally to specify a default set of parameters for all sites. Each site
can now either be configured to use the global defaults or set it's own
parameters.

For most environments it is sufficient to configure the global defaults
and make all sites use the defaults.

Using the global defaults makes it a lot easier to maintain the settings
since you don't need to configure the settings for each site individually
anymore.

Change-Id: I718aa6ccc6c5a7c4044f304f808c08178f8ac2e7