5068 Add error message to report in case some sites are not reachable
When you create a report in a distributed environment with multiple sites
and not all of the sites are reachable when the report is being created
then there was simply missing data without any further notice.

No at the end of the report there is added an error message if there
was a problem with at least one site. That way the reader knows that
the data might be incomplete.

Change-Id: I78704f1f935f7d3e46a1066a19c81518289f70fc