5041 FIX Removed NAGVIS_URLS OMD config option
The "omd config" option NAGVIS_URL which could be used to make the NagVis icons
link to different GUIs has been removed because all CMK environments use the CMK
GUI. A hook for this is not needed.

In case you need to customize your NagVis URLs you can still do it with the
NagVis configuration.

Change-Id: I2090b5b2bc2a583ed8ace4591d8eb7c87b598e81