WinAgent: Call WinAPI through adaptor interface
The target is to clean dependencies to Win API and to place them behind
an adaptor interface. This is necessary for unit testing, among others.
Use basic dependency injection: pass reference to adaptor instance
instead of calling windows API directly.

Additional fixes for portability and mocking:

* Use int64_t and uint64_t instead of Windows-specific (unsigned) __int64.
* Some win_error.* preprocessor _WIN32 defines were broken: they did not
compile without MINGW (that apparently always had _WIN32 on).
* CreateService from WinAPI has 13 (!) parameters, anything above 10
cannot be properly mocked with Google Mock. We never use the last 5
parameters (always zeroes), so they could be left out of WinApiAdaptor
and be filled in at the last step.

Additionally, refactor Environment in order to clear valgrind errors:

 * Use const string members that are initialized in constructor member
   initializer list.
 * Use std::vector<char> as char buffer instead of manually
   heap-allocated buffer.

Change-Id: I6a9503ab837c900d1f928c9d86cc711869c2e139