5047 FIX EC views: Fixed "Events of host X" views of unrelated events
When clicking on the host name of an unrelated event (one that has no host in the monitoring core)
the target page was not showing any event since Check_MK 1.4. The events are now correctly filtered
by the original event host instead of the host name of the found core host.

Change-Id: Ib61ae1c3f955845aa0fdfde1ac8c272717605707