4802 mk_jolokia: Can now be configured to use the local machine's FQDN to access Jolokia
Previously, the Jolokia server could be configured via an IP address or an explicit FQDN.
When using HTTPS on a large number of devices, configuring "localhost" or 127.0.0.1
would lead to certificate errors unless one configured the FQDN for each host individually.
This is unreasonably tedious, and can be avoided using this feature.

Change-Id: I786ba5170f4cd8e06cda8fa2191a0a49c26c289b