5082 FIX gude_powerbanks: Adapted discovery of powerbanks
Devices with OID_END=38 are 12 port power switches with two powerbanks.
Means each powerbank has 6 outlets. Here we can use ChanStatus in order
to find out if one powerbank is enabled/used.

Devices with OID_END=19 are simple switch outlets: 1 Port and 1 powerbank
Once it's plugged in, the state is "on". Thus we use PortState in
discovering function.

Change-Id: I7311aca83abd3ae055e466114b43f7eab44e0d4e