4813 All recipents of an email are visible in notifications
Notification multiplexing can now be used for HTML and ASCII emails.
By default it is enabled on new and existing Check_MK installations.
If notification multiplexing is enabled only one notification is generated
even if multiple recipients are going to be notified. Therefore, a
user can see and reply to all recipients of a notification.
Furthermore, the notification system will be more performant if graphs
are present in the notification the since only one notification is
generated. The notification multiplexing can be disabled in the section
Notification Method of a notification rule.

Change-Id: I5b0d6a59503c925cecb890787527b78a8db69f8f