5057 FIX Fixed cancelling events by "syslog application" when using similar rules
When using two or more rules that both have a "match syslog application" condition and a
"syslog application to cancel event" condition while all other conditions are equal, it
could happen that the first rule always caught messages that should not be caught by that
rule. These messages were not handled by the second rule as expected.

Change-Id: Icdb5a02711ad43e411f89820c8bc3dc4d91087fd