5188 Sending service metrics to Graphite/InfluxDB can now be fine-tuned
Previously, there was a service rule to configure one of three levels of
information to be sent to Graphite/InfluxDB: Send nothing, send only value,
or send value/warn/crit/min/max.

Now the rule is more fine-grained, so one can choose for each part of the
metric if it should be sent or not.  Furthermore, one can now send the unity
of measurement, too.

Change-Id: Ibc75ed41a256768e5f46674aff15f2be75db4aab