5141 WATO Web API: Introduced parameter request_format
Web-API calls which set/update specific elements in WATO had trouble interpreting the incoming data.

JSON encoded data is not fully capable to represent the python syntax - in detail tuples, it uses lists instead.
WATO raises an error when a list is reported instead of a tuple.
For example, this error may happen when you set a SNMPv3 community for a specific host, since the community is
configured as a tuple.

This werk introduces the <tt>request_format=python</tt> parameter.
If it is set, the request is decoded as python instead of json.

Change-Id: I3de66d530861853d033dea33991916d0f6980e28