5203 FIX Removed sorting of perfometers
The user could interactively sort a view by the sorter by clicking on the column header. The
sorter was not sorting by the shown value(s), but was simply using the first perfdata value
of the service for sorting. This could be wrong since the base for the shown value is often
not the first perfdata value.

The broken sorting function has been removed now.

Maybe we could add a more intelligent sorting in the future. For the moment you can create
a custom view and add a sorter that sorts the view using a custom perfdata value.

Change-Id: I855f873bcf597d6c97f62b7c1fbda248b57c5d6d