5204 FIX Host rename: Fixed broken rename of hosts in rrdcached journal
When using the new Check_MK rrd format (single value per RRD), the
host rename feature was not renaming host specific entries in the
journal of the rrdcached.

This resulted in lost RRD values that have not been written to the
RRDs yet.

Manual fix: You could stop the rrdcached (or the whole site), go
to var/rrdcached and rename the host in the journal files manually,
e.g. using <tt>sed</tt>.

Change-Id: Ib330541438a87692f40a2856add7d1f066379c01