5259 FIX cmk-update-agent: Improving robustness in curl handling
In some cases, cmk-update-agent might abort showing the following Error:
AttributeError: 'NoneType' object has no attribute 'group'

This Error only occurs when curl is used for server communications. This is the case if python-requests is not installed.

After updating Check_MK and rebaking your agents, it is also required to reinstall the Check_MK Agent on affected hosts.
A workaround would be to install the Python requests module on the affected hosts.

Change-Id: If69c4bf1b79d0a8496ad0b4415c4c5aa7e92167c