5017 FIX check_mk_agent: Fixed false positive error messages while installing rpm from bakery
Previously a rpm package of the linux agent assumed sysvinit while checking for
xinetd to be installed. The check failed if systemd was the init system. This
affects only rpm packages built by agent bakery.

Now the rpm packages verifies the presence of xinetd by looking for the
presence of the binary (/usr/sbin/xinetd) which seems to be less error-prone.

Change-Id: I03b2ed41d5da7d1e6c49e9c501a3d5971a263a59