5034 FIX Do not skip first entry in Windows eventlog
The eventlog reference for the first record to be read was wrongly
advanced one past the previously read entry already when seeking
for new entries. This caused the reference to point past the actual
end of eventlog and skip the first message that arrived after that.
The correct behaviour is to only advance the reference if there really
are new entries in the eventlog and not only when trying to seek one
past the previous end.

Change-Id: Ie6ce104b6484cdc1bd1bce56a3f158a7c39eb4f3