Fix Windows Agent crash with vista_api = yes
The crash was caused by refactoring the dynamically loaded functions
in EvtFunctionMap, done in commit
16d8a0f1affed78bbe910dcb62dc73c06ee14508. The original reason for the
problematic change was to try to avoid Windows headers in unit tess.
Now that we can use Windows headers in tests through Wine, the
problematic change with dynamically loaded functions in EvtFunctionMap
can be simply reverted.

Change-Id: I4348886989c4f704a5c26c562da7b8b55db9d22b