Fix Windows Agent crash with vista_api = yes
The crash was caused by refactoring the dynamically loaded functions
in EvtFunctionMap, done in commit
858d140b076d4c236996f5f26c374c70d38943d7. The original reason for the
problematic change was to try to avoid Windows headers in unit tess.
Now that we can use Windows headers in tests through Wine, the
problematic change with dynamically loaded functions in EvtFunctionMap
can be simply reverted.

Change-Id: I4348886989c4f704a5c26c562da7b8b55db9d22b