Fix Windows eventlog output with vista_api=yes
The Vista API uses the Win API function EvtFormatMessage that delivers
formatted strings with trailing null characters within the allocated
wstring buffer size. Until the refactoring done in
37876832066f90a61c3dd0c9c333cb719c2af135, this did not cause trouble as
the wstring to UTF8 conversion was done using the Win API and 'old' C
strings. The refactoring introduced C++ standard library conversion that
is done on wstring objects instead of C strings. As a consequence, the
trailing null character was taken into the formatted message, later
causing the socket output to be cut at it. The fix is to trim possible
trailing null characters away from the formatted eventlog message.

Change-Id: I2bf5c8093c79e6e7babcc6abb15b91e6d23756d5