5211 FIX Agent bakery: Reduced agent baking time when monitoring 0 hosts from central site
When having a distributed Check_MK/WATO environment where the central management site
has zero hosts configured to be monitored, the time needed to bake agents was increased
by a broken cache. This cache is now being used even when the central site is not monitoring
any host.

You could workaround this issue by adding a single host to be monitored from the central site.

Change-Id: I448f69dcf59d5469f0aaf5c0f9457c34d8b49cfd