Fixed wrong cache usage
An empty cache does not mean that it has not be populated. An empty
population result can also be a regular case in which the calculation
should not be done twice.

Example: A sites needs to detect the list of hosts to be monitored
locally. Even when there are hundret of sites with thousands of hosts
monitored on all sites, a empty result has to be cached for a site
that does not monitor any host.

Change-Id: Id69612525c3e8ffc3b02e578f8a4543aad055cc3