5216 FIX Failed notification checking is now limited to 7 days
This change fixes a GUI (sidebar) rendering performance issue which
was caused by the check for failed notifications. For individual
admin users the GUI could be very slow because the core was searching
far in the past of the monitoring history.

The failed notification checking is now being limited by a configurable
"Failed notification horizon" which is set to 7 days be default. You can
configure this value according to your needs in the Global settings.

Change-Id: If372bc9560343432c00376a3f983ec1df68c0a33