windows agent chroot can now be acessed in parallel
When CHROOT_BUILD_DIR is set to a custom directory name when
calling the make-agent script, this will be used as mount point
to mount the current Check_MK base path into.

This makes it possible to use a single chroot for multiple
build jobs in parallel. This is needed for our build system.

Change-Id: Ibfe6a82041d8dede30d375347f81ad49aa19ad87