5362 FIX Fixed possible exception on global settings page related to "diskspace cleanup"
In 1.2.8 or older it was possible to configure "max_file_age = None" in the etc/diskspace.conf
file. This value was not interpreted correctly by the current code which lead to an exception
"TypeError: unsupported operand type(s) for divmod(): 'NoneType' and 'int'" on the global settings
page.

To workaround this issue, you could remove this option form the etc/diskspace.conf file.

Change-Id: If6ed76861a6596d034af4209b2c6765b2ca533ea