Use GNU autotools for building Windows agent
The Makefile for the Windows agent has been a source of many problems,
e.g. in its manual defition of build dependencies that
has been prone to errors and preventing e..g efficient build
parallelization. With GNU autoconf and automake the shortcomings of a
manually maintained Makefile can be overcome, immediately cutting build
times (-50% for a full build, even more for incremental builds due to
proper handling of dependencies) and hopefully making the build process
easier to maintain.

The 32 and 64 bit builds for Windows are now configured with different
parameters passed to the auto-generated configure script. The builds
are then run in parallel, with the possibility of specifying N parallel
jobs with the option '-jN' to the build scripts.

The README and CONTENTS files are brought up to date.

Change-Id: I8495d0ac3a47bd45d4193a4e0763fc202dd84bfe