Remove dependency on Git for Windows agent build
Windows agent has had a dependency to Git revision that is included as
extra information for crash debug log prints. As Git is not available
during the build on our build server, this dependency cannot be there
any more. Just give the Check_MK version out in the crah log, instead.

Change-Id: Ie9f0b0ee5bd7e44e98160a41bdc8d77f0f3105c9