5341 agent_vsphere: change of the option "Display VM power state on"
The option "Display VM power state on" of the WATO rule now displays
the power state of VMs <i>additionally</i> on either the ESX host or
the VM. The option is therefore renamed to "Display power state
additionally on". This will result in services for <i>both</i> the
queried system and the ESX host / VM. By disabling the unwanted
services it is then possible to configure where the corresponding
services are displayed. Because the data is now still available on
the queried system the service "Object count" will now display the
correct count when the option is used.

Change-Id: I89172eaa9384a7ef0436a1084ce812ae9cdd2d07