Windows agent: add integration tests
Add pytest-based integration tests. The purpose is to offer basic
regression tests to ensure that basic functionality remains intact
across refactorings etc.

The test setup requires a local Linux machine to run the tests on. The
user at the Linux machine must have ssh RSA key pair specifically
generated for this purpose as ~/.ssh/id_rsa_windows_test[.pub]. A
dedicated Windows server must be available for remote access over SSH
with the public key in its authorized keys.

Each pytest test case will be initiated on local Linux machine but
switch over SSH to the remote Windows server and run the actual test
case there. Per-session setup and teardown, e.g. copying files over SSH
is done with the help of per-session pytest fixtures defined in
conftest.py.

The tests cover most of the sections in the output of the Windows
agent. Currently excluded are the sections MS Exchange and Skype as
these require the installations of these commercial software products.
The output is verified against expected patterns and is likely to find
the vast majority of missing or malformatted output. Substantial errors
resulting from broken functionality but that are formatted and output
correctly are likely to go unnoticed.

The sections 'plugins' and 'local' are tested with the plugins
netstat_an.bat and wmic_if.bat. The target is to ensure the correct
functionality of the plugin execution framework of the Windows agent.
The tests can be later extended to test the execution of other plugins,
if necessary.

Besides the output of agent sections during normal execution, the start
parameters of the Windows agent are tested. These tests would interfere
with possibly existing Check_MK_Agent service on the system. This is why
a possibly existing installation is removed during test setup.

Inter-process synchronized access to lock file is used in order to
prevent simultaneous execution of tests in several processes.

PENDING CHANGES:
The build target integrationtest will be committed separately later when
the build server(s) support the test-specific SSH keys. Before that, a
premature commit of the build changes would break the build.

Change-Id: I26e5543924e07acff56be16f9fe6724a8f78358a