Overhauled ListColumn handling.
The previous isEmpty() and makeContains() methods were not flexible enough
to handle upcoming features like regexp matching, so they were replaced by a
more general mechanism: Every ListColumn now knows how to calculate their
contents as a stream of strings.  As a first approximation, we simply use a
vector for this stream, we can later improve performance by replacing that
with a range or iterators.

Several bugs regarding AuthUser were fixed on the way: Often the AuthUser
was just used for rendering a column, but omitted for filtering.

Change-Id: I773150c8e6e7f5c29951d762314b56eb504b766c