5402 Number of apache processes can now be configured via Global Settings
It is now possible to configure the apache configuration option {MaxClients} via
WATO (Global Settings > Site Management > Apache process tuning).

This option is often changed to tune the site apache to either consume less memory
(reduce the number of maximum processes) or being able to handle more users in the
GUI simultaneously.

The option can be configured globally for all sites or in the site specific globals,
depending on your needs.

The defaul value of this option is read from the existing configuration files
(etc/apache/apache.conf and etc/apache/conf.d/*.conf) and defaults to {64}.

Change-Id: I1cd97e6a3391d2cb5708a8a69fe9415e0a362cb9