5405 FIX LDAP: Removed missleading and confusing "No persistent connection" option
The LDAP connector had an option to tell it that the connection to the LDAP server
should not be reused. This made the connector reconnect to the LDAP server for each
LDAP query.

Because enabling this option can result in a much longer synchronization run, the
option has now been dropped.

Change-Id: Ibf7524bc3038979cc8c67a29f468bac13eb2fafe