Make RRDColumn a subclass of ListColumn.
This enables filtering on this column, but more importantly: It unifies
things quite a bit.

As a side effect, we have to thread through the time zone offset in a few
places, but that's somehow nice, too: It makes the offset handling more
explicit.  We simplify and clean this up later.

Change-Id: I98a5b27abd1c5c9712ffb4d45aaa0747eca2eeb1