5431 SEC Fixed possible reflected XSS using custom bookmarks
It was possible to create custom bookmarks by making the user open a
crafted URL. This created a bookmark in the users default bookmark list
which could be used to execute custom javascript code when the user
clicks on the just created link.

For example the user session cookies can be read and reported to the
attackers, who could then hijack the users sessions with the application.

This issue has been fixed by limiting absolute URLs in bookmarks to the
URL schemes <tt>https</tt> and <tt>http</tt>.

Change-Id: I80a47c981970f23dd9737f1c6000cc82799220b1