5196 FIX Bumped Livestatus stack size to give regex matching more breathing room.
The current implementation of regular expression matching can use a
non-trivial amount of memory (mainly stack), and our previous memory limit
was a bit tight.  This has been improved, so regular expression matching
should work more reliably.

Note that this increases the overall amount of memory allocated for
Check_MK, but for normal installations the increase should be negligible.

Change-Id: Iff55bc95ba8a316174a4830a68939b5a16be04df