5354 FIX cmctc_lcp: Fixes an error leading to wrong inventory and check output
This werk fixes an error which caused that the inventory for some sensors
was not performed correctly. This resulted in a wrong amount of discovered
services and a wrong output (numbers and units) for these services .
Therefore, the services of the corresponding hosts have to be rediscovered.
Temperature sensors are not affected by this error. One other type of
sensor is not affected as well. Which sensor type it is might depend on the
used Python version. In tests using Python version 2.7.14 Blower sensors
were not affected.

Change-Id: I4c4d941006a2ea19acd42582359fd3f035b15a9c