5417 FIX Windows: allow whitespace in eventlog event source
In version 1.2.8, whitespace in event source name produced the error
message "The system cannot find the path specified." The reason was
premature formatting of the event source name for output by replacing
spaces with underscore, which led to failure when trying to find the
DLL for translating the eventlog message. The bug is no longer present
in version 1.4.0.

Change-Id: I37a5b5c5ffc6f9ce58ec613afaad7e7f3f081baf