5436 FIX Fixed logical problem with SNMP check interval rule
The rule was previously configured for each individual check. This was not really
correct, because the SNMP data is fetched for main checks and their sub checks
together. This means that you can not define inidividual SNMP checks intervals for
these checks.

The ruleset "Check intervals for SNMP checks" has now been changed to work with the
"section names" instead of inidividual checks.

Change-Id: I56f00bd5e49f4bf791801da94740b0329eb83671