5518 FIX win.mem: Fix incorrect representation of absolute memory levels in graphs
Previously, a warn or crit level of eg. 20 GB would be displayed in the graph as 20 TB
due to the levels not being set to the same unit as the performance value itself.
The problem only affected the levels passed to the metrics system for representation,
not the actual levels used for evaluation.

Change-Id: Id9845a02f886226d58f13a1734613f7b5f5f9b8d